Adopt to v0.28.0 (#121)

+++ api/models/v1_postgres_backup_config_create_request.go
@@ -36,7 +36,7 @@ type V1PostgresBackupConfigCreateRequest struct {
 	ProjectID string `json:"projectID,omitempty"`
 
 	// retention
-	Retention int32 `json:"retention,omitempty"`
+	Retention int64 `json:"retention,omitempty"`
 
 	// s3 bucket name
 	S3BucketName string `json:"s3BucketName,omitempty"`
